add recipe building for OG-USA Tax-Calc like B-Tax

+++ pre_commit_script.py
@@ -6,19 +6,34 @@ import tempfile
 from urllib.request import urlopen
 import hashlib
 
-PACKAGES = ('http://github.com/PeterDSteinberg/B-Tax',)
+PACKAGES = ('B-Tax',
+            'Tax-Calculator',
+            'OG-USA',)
 TOP_DIR = os.path.dirname(__file__)
 RECIPES_DIR = os.path.join(TOP_DIR, 'recipes')
 
-META_VARIABLES = {'B-Tax': {
-                    'name': "btax",
-                    'repo_name': "B-Tax",
-                    'git_org': "PeterDSteinberg",
-                    'version':  "0.1.091",
-}}
+META_VARIABLES = {
+    'B-Tax': {
+        'name': "btax",
+        'repo_name': "B-Tax",
+        'rel_recipe': 'conda.recipe',
+    },
+    'Tax-Calculator': {
+        'name': 'taxcalc',
+        'repo_name': 'Tax-Calculator',
+        'rel_recipe': 'conda.recipe',
+    },
+    'OG-USA': {
+        'name': 'ogusa',
+        'repo_name': 'OG-USA',
+        'rel_recipe': os.path.join('Python', 'conda.recipe'),
+    }
+}
 
-def rm_recipes_dir_contents():
+def rm_recipes_dir_contents(package):
     for item in os.listdir(RECIPES_DIR):
+        if not package in item:
+            continue
         item = os.path.join(RECIPES_DIR, item)
         if os.path.isfile(item):
             os.remove(item)
@@ -39,54 +54,74 @@ def proc_wrapper(args, cwd):
         raise ValueError('Subprocess has bad return code {}'.format(proc.poll()))
 
 
-def get_latest_packages(btax_branch, btax_version):
+def get_latest_packages(btax_branch=None, btax_version=None,
+                        taxcalc_version=None, taxcalc_branch=None,
+                        ogusa_version=None, ogusa_branch=None,
+                        git_org='open-source-economics'):
     cmd = ['git', 'clone']
     for package in PACKAGES:
-        if 'B-Tax' in package:
-            meta_var = META_VARIABLES['B-Tax']
-            meta_var['version'] = btax_version
-            source_url = 'https://github.com/{git_org}/{repo_name}/archive/{version}.tar.gz'
-            source_url = source_url.format(**meta_var)
-            meta_var['source_url'] = source_url
+        if package == 'B-Tax':
+            if not btax_version:
+                print('Not building B-Tax')
+                continue
+            else:
+                version = btax_version
+        elif package == 'Tax-Calculator':
+            if not taxcalc_version:
+                print('Not building Tax-Calculator')
+                continue
+            else:
+                version = taxcalc_version
+        elif package == 'OG-USA':
+            if not ogusa_version:
+                print('Not building OG-USA')
+                continue
+            else:
+                version = ogusa_version
+        rm_recipes_dir_contents(package)
+        meta_var = META_VARIABLES[package]
+        meta_var['git_org'] = git_org
+        meta_var['version'] = version
+        source_url = 'https://github.com/{git_org}/{repo_name}/archive/{version}.tar.gz'
+        source_url = source_url.format(**meta_var)
+        meta_var['source_url'] = source_url
 
-            print(source_url)
-            with urlopen(source_url) as f:
-                tar_file = f.read()
-                meta_var['sha256'] = hashlib.sha256(tar_file).hexdigest()
-        else:
-            raise NotImplementedError("Add more keys to the META_VARIABLES global")
-        args = cmd + [package]
-
+        print(source_url)
         tmp = tempfile.mkdtemp()
         try:
-            proc_wrapper(args, tmp)
-            pdir = package.split('/')[-1]
-            repo_clone = os.path.join(tmp, pdir)
-            proc_wrapper(['git', 'fetch', '--all'], repo_clone)
-            proc_wrapper(['git', 'checkout', btax_branch], repo_clone)
-            recipe = os.path.join(repo_clone, 'conda.recipe')
-            shutil.copytree(recipe, os.path.join(RECIPES_DIR, pdir))
-            meta_yaml = os.path.join(RECIPES_DIR, pdir, 'meta.yaml')
-            with open(meta_yaml) as f:
-                contents = f.read()
-            new_lines = []
-            for k, v in meta_var.items():
-                line = '% set {0} = "{1}" %'.format(k, v)
-                new_lines.append('{' + line + '}')
-            contents = "\n".join(new_lines) + '\n' + contents
-            with open(meta_yaml, 'w') as f:
-                f.write(contents)
+            with urlopen(source_url) as f:
+                tar_file_contents = f.read()
+                meta_var['sha256'] = hashlib.sha256(tar_file_contents).hexdigest()
+                tar_file = os.path.join(tmp, '{version}.tar.gz'.format(**meta_var))
+                with open(tar_file, 'wb') as f:
+                    f.write(tar_file_contents)
+                proc_wrapper(['tar', '-xvf', tar_file], cwd=tmp)
+                repo_clone = os.path.join(tmp, '{repo_name}-{version}'.format(**meta_var))
+                recipe = os.path.join(repo_clone, meta_var['rel_recipe'])
+                local_recipe = os.path.join(RECIPES_DIR, package)
+                shutil.copytree(recipe, local_recipe)
+                meta_yaml = os.path.join(local_recipe, 'meta.yaml')
+                with open(meta_yaml) as f:
+                    contents = [x for x in f.readlines() if not x.lstrip().startswith('#')]
+                new_lines = []
+                for k, v in meta_var.items():
+                    line = '% set {0} = "{1}" %'.format(k, v)
+                    new_lines.append('{' + line + '}')
+                contents = "\n".join(new_lines) + '\n' + '\n'.join(contents)
+                with open(meta_yaml, 'w') as f:
+                    f.write(contents)
         finally:
             if os.path.exists(tmp):
                 shutil.rmtree(tmp)
 
 def pre_commit_main():
     parser = argparse.ArgumentParser(description='Build conda-forge packages')
-    parser.add_argument('--btax-branch', required=True)
+    parser.add_argument('--git-org', required=True)
     parser.add_argument('--btax-version', required=True)
+    parser.add_argument('--taxcalc-version', required=True)
+    parser.add_argument('--ogusa-version', required=True)
     args = parser.parse_args()
-    rm_recipes_dir_contents()
-    get_latest_packages(args.btax_branch, args.btax_version)
+    get_latest_packages(**vars(args))
 
 
 if __name__ == "__main__":
